refactor: improve type safety in response validation and workspace API tests
